+++ lib/src/screenshot_service.dart
@@ -1,20 +1,36 @@
 import 'dart:io';
+import 'dart:math' as math;
 import 'dart:ui' as ui;
 
 import 'package:colaxy_screenshot/colaxy_screenshot.dart';
 import 'package:device_frame_plus/device_frame_plus.dart';
 import 'package:easy_localization/easy_localization.dart';
-import 'package:flutter/material.dart' hide Image;
+import 'package:flutter/material.dart';
 import 'package:flutter/rendering.dart';
 import 'package:flutter/services.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
-import 'package:image/image.dart';
+import 'package:image/image.dart' hide Color, Image;
+import 'package:window_size/window_size.dart';
 
 /// Android用のロケールマッピング
-const _androidLocaleMap = {'en': 'en-US', 'ja': 'ja-JP', 'zh': 'zh-CN'};
+const _androidLocaleMap = {
+  'en': 'en-US',
+  'ja': 'ja-JP',
+  'zh': 'zh-CN',
+  'es': 'es-ES',
+  'pt': 'pt-PT',
+  'tr': 'tr-TR',
+};
 
 /// iOS用のロケールマッピング
-const _iOSLocaleMap = {'en': 'en-US', 'ja': 'ja', 'zh': 'zh-Hans'};
+const _iOSLocaleMap = {
+  'en': 'en-US',
+  'ja': 'ja',
+  'zh': 'zh-Hans',
+  'es': 'es-ES',
+  'pt': 'pt-PT',
+  'tr': 'tr',
+};
 
 /// メインのスクリーンショットサービス
 class ScreenshotService {
@@ -28,6 +44,9 @@ class ScreenshotService {
 
   /// スクリーンショットを実行する
   Future<void> executeScreenshots() async {
+    // set Feature Graphic Page
+    await getFeatureGraphicScreenshot();
+
     final defaultDelay = config.captureDelay;
     var isFirst = true;
     // 各デバイス × 各言語 × 各ページの組み合わせでスクリーンショットを作成
@@ -50,6 +69,106 @@ class ScreenshotService {
     }
   }
 
+  /// Feature Graphic Page generate
+  Future<void> getFeatureGraphicScreenshot() async {
+    _appKey = GlobalKey();
+
+    final Widget appWidget = ProviderScope(
+      overrides: config.overrides,
+      child: config.easyLocalizationWrapper(
+        Builder(
+          builder: (context) {
+            return FutureBuilder(
+              future: () async {
+                Intl.defaultLocale = 'en';
+                await context.setLocale(const Locale('en', 'US'));
+                return null;
+              }(),
+              builder: (_, __) => config.wrapFunction(
+                Container(
+                  decoration: const BoxDecoration(
+                    gradient: LinearGradient(
+                      begin: Alignment.topLeft,
+                      end: Alignment.bottomRight,
+                      colors: [
+                        Color(0xFF8BC34A), // Light Green
+                        Color(0xFF009688), // Teal
+                      ],
+                    ),
+                  ),
+                  child: Scaffold(
+                    backgroundColor: Colors.transparent,
+                    body: Stack(
+                      children: [
+                        Positioned(
+                          top: 0,
+                          left: 200,
+                          child: Transform.rotate(
+                            angle: -math.pi / 6, // 数字が大きければ回転角度が小さくなる
+                            child: ClipRRect(
+                              borderRadius: BorderRadius.circular(60),
+                              child: Image.asset(
+                                'assets/app_icons/icon.png',
+                                width: 400,
+                                height: 400,
+                                fit: BoxFit.cover,
+                              ),
+                            ),
+                          ),
+                        ),
+                        Positioned(
+                          top: -50,
+                          right: 200,
+                          child: Transform.rotate(
+                            angle: math.pi / 6, // 数字が大きければ回転角度が小さくなる
+                            child: DeviceFrame(
+                              device: const ScreenshotModeInfo(
+                                mode: ScreenshotMode.phone,
+                                deviceSize: Size(642, 1389),
+                              ).toDeviceInfo(),
+                              screen: config.featureGraphicPage,
+                            ),
+                          ),
+                        ),
+                      ],
+                    ),
+                  ),
+                ),
+              ),
+            );
+          },
+        ),
+      ),
+    );
+    setWindowToSize(const Size(1024, 500));
+    runApp(RepaintBoundary(
+      key: _appKey,
+      child: appWidget,
+    ));
+    await Future<void>.delayed(const Duration(seconds: 3));
+    await WidgetsBinding.instance.endOfFrame;
+    final boundary =
+        _appKey?.currentContext!.findRenderObject() as RenderRepaintBoundary?;
+    final image = await boundary!.toImage(pixelRatio: 3);
+    final byteData = await image.toByteData(format: ui.ImageByteFormat.png);
+    final imageBytes = byteData?.buffer.asUint8List();
+    if (imageBytes == null) {
+      throw Exception('スクリーンショットの撮影に失敗しました');
+    }
+    final pngImage = decodePng(imageBytes)!;
+    final resizedImage = copyResize(pngImage, width: 1024, height: 500);
+    File('$appPath/fastlane/metadata/android/featureGraphic.png')
+        .writeAsBytesSync(encodePng(resizedImage));
+  }
+
+  void setWindowToSize(Size deviceSize) {
+    const rate = 3.3;
+    setWindowMinSize(deviceSize / rate);
+    setWindowMaxSize(deviceSize / rate);
+    setWindowFrame(
+        Rect.fromLTWH(100, 100, deviceSize.width, deviceSize.height));
+  }
+
   /// 言語設定を含むアプリウィジェットを構築
   Widget _buildAppWithLocale({
     required Locale locale,
@@ -63,15 +182,8 @@ class ScreenshotService {
         ...config.overrides,
         ...page.overrides ?? [],
       ],
-      child: EasyLocalization(
-        supportedLocales: const [
-          Locale('ja', 'JP'),
-          Locale('en', 'US'),
-          Locale('zh', 'CN'),
-        ],
-        path: 'assets/translations',
-        fallbackLocale: const Locale('en', 'US'),
-        child: Builder(
+      child: config.easyLocalizationWrapper(
+        Builder(
           builder: (context) {
             return FutureBuilder(
               future: () async {