+++ CHANGELOG.md
@@ -1,3 +1,15 @@
+## 0.4.0
+
+- Fix some issues.
+
+## 0.3.0
+
+- Add easyLocalizationWrapper to config.
+
+## 0.2.0
+
+- Add support for Spanish, Portuguese, and Turkish.
+
 ## 0.1.1
 
 - Add screenshotId to screenshot file name.